--- plugins/faustwp/phpunit-multisite.xml.dist
@@ -1,30 +1,0 @@
-<?xml version="1.0"?>
-<phpunit
-  backupGlobals="false"
-  bootstrap="tests/bootstrap.php"
-  colors="true"
-  convertErrorsToExceptions="true"
-  convertNoticesToExceptions="true"
-  convertWarningsToExceptions="true"
-  testdox="true"
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:noNamespaceSchemaLocation="https://schema.phpunit.de/9.3/phpunit.xsd"
->
-  <php>
-    <const name="WP_TESTS_MULTISITE" value="1" />
-  </php>
-  <coverage processUncoveredFiles="true">
-    <include>
-      <directory suffix=".php">./includes</directory>
-    </include>
-    <report>
-      <clover outputFile="coverage/clover.xml"/>
-    </report>
-  </coverage>
-  <testsuites>
-    <testsuite name="integration">
-      <directory prefix="test-" suffix=".php">./tests/integration/</directory>
-    </testsuite>
-  </testsuites>
-  <logging/>
-</phpunit>
