+++ plugins/faustwp/phpunit.xml.dist
@@ -19,8 +19,11 @@
     </report>
   </coverage>
   <testsuites>
+    <testsuite name="unit">
+      <directory suffix=".php">./tests/unit/</directory>
+    </testsuite>
     <testsuite name="integration">
-      <directory prefix="test-" suffix=".php">./tests/integration/</directory>
+      <directory suffix=".php">./tests/integration/</directory>
     </testsuite>
   </testsuites>
   <logging/>
