+++ plugins/faustwp/tests/intelephense-alias.php
